+++ packages/tokens/scripts/formats/css-vars.js
@@ -126,10 +126,6 @@ export const cssVariableFormatter = {
       // but dont want it to appear. We can remove it later, but for now we want to keep it in the codebase to not break the header.
       'color-shadow-sticky-down-color',
 
-      // #1716: This token is not available anymore in figma, but the API still fetches it.
-      // Kind of a zombie token. As I cannot get rid of it, I will ignore it for now. We can remove it later when we clean up the codebase.
-      'range-color-readonly',
-
       // #1171: Leftover after refactoring the readonly tokens.
       // Can be removed as soon as it is gone in Figma.
       'readonly-opacity-color',